Adjusted project structure for better management. Created basic kotlin class that will contain methods for database communication.
