+++ app/src/main/java/com/example/parkmanager/activities/LoginActivity.kt
@@ -1,4 +1,4 @@
-package com.example.parkmanager
+package com.example.parkmanager.activities
 
 import android.os.Bundle
 import android.os.PersistableBundle
@@ -28,21 +28,19 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 
-class RegisterActivity : ComponentActivity() {
+class LoginActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?, persistentState: PersistableBundle?) {
         super.onCreate(savedInstanceState, persistentState)
-        setContent { RegisterScreen {} }
+        setContent { LoginScreen {} }
     }
 }
 
 /**
- * User registration screen.
+ * User login screen.
  */
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun RegisterScreen(onRegistrationSuccess: () -> Unit) {
-    var name by remember { mutableStateOf("") }
-    var surname by remember { mutableStateOf("") }
+fun LoginScreen(onLoginSuccess: () -> Unit) {
     var email by remember { mutableStateOf("") }
     var password by remember { mutableStateOf("") }
 
@@ -53,24 +51,9 @@ fun RegisterScreen(onRegistrationSuccess: () -> Unit) {
         verticalArrangement = Arrangement.Center,
         horizontalAlignment = Alignment.CenterHorizontally
     ) {
-        Text("Registration Screen", fontSize = 20.sp, fontWeight = FontWeight.Bold)
+        Text("Login Screen", fontSize = 20.sp, fontWeight = FontWeight.Bold)
+
         Spacer(modifier = Modifier.height(16.dp))
-
-        OutlinedTextField(
-            value = name,
-            onValueChange = { name = it },
-            label = { Text("Name") }
-        )
-
-        Spacer(modifier = Modifier.height(8.dp))
-
-        OutlinedTextField(
-            value = surname,
-            onValueChange = { surname = it },
-            label = { Text("Surname") }
-        )
-
-        Spacer(modifier = Modifier.height(8.dp))
 
         OutlinedTextField(
             value = email,
@@ -87,46 +70,29 @@ fun RegisterScreen(onRegistrationSuccess: () -> Unit) {
             visualTransformation = PasswordVisualTransformation()
         )
 
-        Spacer(modifier = Modifier.height(8.dp))
-
-        OutlinedTextField(
-            value = password,
-            onValueChange = { password = it },
-            label = { Text("Confirm Password") },
-            visualTransformation = PasswordVisualTransformation()
-        )
-
         Spacer(modifier = Modifier.height(16.dp))
 
         Button(
             onClick = {
-                performRegistration(
-                    name,
-                    surname,
+                performLogin(
                     email,
                     password,
-                    onRegistrationSuccess
+                    onLoginSuccess
                 )
             },
             modifier = Modifier.fillMaxWidth()
         ) {
-            Text("Register")
+            Text("Login")
         }
     }
 }
 
 @Preview(showBackground = true)
 @Composable
-fun RegisterScreenPreview() {
-    RegisterScreen {}
+fun LoginScreenPreview() {
+    LoginScreen {}
 }
 
-fun performRegistration(
-    name: String,
-    surname: String,
-    email: String,
-    password: String,
-    onRegistrationSuccess: () -> Unit
-) {
+fun performLogin(email: String, password: String, onLoginSuccess: () -> Unit) {
     TODO("Not yet implemented")
 }
