+++ app/src/main/java/com/example/parkmanager/activities/NewReservationActivity.kt
@@ -1,4 +1,4 @@
-package com.example.parkmanager
+package com.example.parkmanager.activities
 
 import android.os.Bundle
 import android.os.PersistableBundle
@@ -8,10 +8,8 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Button
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -21,18 +19,15 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 
-class AdminScreenActivity : ComponentActivity() {
+class NewReservationActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?, persistentState: PersistableBundle?) {
         super.onCreate(savedInstanceState, persistentState)
-        setContent { AdminAccessScreen() }
+        setContent{ NewReservation() }
     }
 }
 
-/**
- * Administrator access screen.
- */
 @Composable
-fun AdminAccessScreen() {
+fun NewReservation() {
     Column(
         modifier = Modifier
             .fillMaxSize()
@@ -40,7 +35,7 @@ fun AdminAccessScreen() {
         verticalArrangement = Arrangement.Center,
         horizontalAlignment = Alignment.CenterHorizontally
     ) {
-        Text("Admin Panel", fontSize = 20.sp, fontWeight = FontWeight.Bold)
+        Text("This activity will handle the process of making a new reservation.", fontSize = 20.sp, fontWeight = FontWeight.Bold)
 
         Spacer(modifier = Modifier.height(16.dp))
     }
@@ -48,6 +43,6 @@ fun AdminAccessScreen() {
 
 @Preview(showBackground = true)
 @Composable
-fun AdminAccessScreenPreview() {
-    AdminAccessScreen()
+fun NewReservationPreview() {
+    NewReservation()
 }
