+++ app/src/main/java/com/example/parkmanager/controllers/DatabaseController.kt
@@ -1,0 +1,5 @@
+package com.example.parkmanager.controllers
+
+class DatabaseController {
+
+}
